Added the custom exception for missing mobile station && changed the id strategy

+++ src/main/java/ee/lagunemine/locatorapi/repository/PositionRecordRepository.java
@@ -3,6 +3,6 @@ package ee.lagunemine.locatorapi.repository;
 import ee.lagunemine.locatorapi.model.PositionRecord;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PositionRecordRepository extends CrudRepository<PositionRecord, Long> {
+public interface PositionRecordRepository extends CrudRepository<PositionRecord, Integer> {
 
 }
