+++ src/main/java/ee/lagunemine/locatorapi/repository/StationMobileRepository.java
@@ -3,6 +3,6 @@ package ee.lagunemine.locatorapi.repository;
 import ee.lagunemine.locatorapi.model.StationMobile;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StationMobileRepository extends CrudRepository<StationMobile, Long> {
+public interface StationMobileRepository extends CrudRepository<StationMobile, Integer> {
 
 }
